test: Only restart KubeDNS if required
Instead of always restarting the kube-dns deployment, split the
validation of the installation into a separate function so we can
validate the deployment and only restart kube-dns if we have to.

Also, parallelize the validation to minimize the time it takes. Longer
term, this will reduce the time per test.

Signed-off-by: Thomas Graf <thomas@cilium.io>